requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
APScheduler==3.10.4
gunicorn>=21.0
